feat: refactor PayloadTab and ResponseTab components for improved payload formatting and add copy functionality, add proper JSON error hanlding of PayloadTab
