+++ packages/vscode-extension/src/network/components/PayloadTab.tsx
@@ -1,36 +1,32 @@
 import { NetworkLog } from "../hooks/useNetworkTracker";
+import IconButton from "../../webview/components/shared/IconButton";
+import { formatJSONBody, formatGETParams } from "../utils/requestFormatUtils";
 
 interface PayloadTabProps {
   networkLog: NetworkLog;
 }
-
-const getParams = (url: string): Record<string, string> => {
-  try {
-    const urlObj = new URL(url);
-    const params: Record<string, string> = {};
-    urlObj.searchParams.forEach((value, key) => {
-      params[key] = value;
-    });
-    return params;
-  } catch {
-    return {};
-  }
-};
 
 const PayloadTab = ({ networkLog }: PayloadTabProps) => {
   if (!networkLog.request) {
     return null;
   }
 
-  const payloadData = JSON.stringify(
+  const payloadData =
     networkLog.request.method === "GET"
-      ? getParams(networkLog.request.url)
-      : JSON.parse(networkLog.request.postData || "{}"),
-    null,
-    2
-  );
+      ? formatGETParams(networkLog.request.url)
+      : formatJSONBody(networkLog.request.postData);
 
-  return <pre>{payloadData}</pre>;
+  return (
+    <>
+      <IconButton
+        className="response-tab-copy-button"
+        tooltip={{ label: "Copy to Clipboard", side: "bottom" }}
+        onClick={() => navigator.clipboard.writeText(payloadData)}>
+        <span className="codicon codicon-copy" />
+      </IconButton>
+      <pre className="response-tab-pre">{payloadData}</pre>
+    </>
+  );
 };
 
 export default PayloadTab;
