+++ packages/vscode-extension/src/network/utils/requestFormatUtils.ts
@@ -1,0 +1,28 @@
+function stringify(obj: unknown): string {
+  return JSON.stringify(obj, null, 2);
+}
+
+export function formatJSONBody(body: unknown): string {
+  if (typeof body !== "string") {
+    return "No response body";
+  }
+  try {
+    const parsed = JSON.parse(body);
+    return stringify(parsed);
+  } catch {
+    return body;
+  }
+}
+
+export function formatGETParams(url: string): string {
+  try {
+    const urlObj = new URL(url);
+    const params: Record<string, string> = {};
+    urlObj.searchParams.forEach((value, key) => {
+      params[key] = value;
+    });
+    return stringify(params);
+  } catch {
+    return stringify({});
+  }
+}
